update docs in the source
