+++ src/utils/fs.rs
@@ -68,7 +68,7 @@ pub fn copy_data_file(src_path: &str, dest_path: &Path) -> Result<(), Box<Error>
 /// I.e. "data/_html-template/css/book.css" will be written to
 /// "assets/css/book.css".
 ///
-/// ```no_run
+/// ```ignore
 /// utils::fs::copy_data("data/_html-template/**/*",
 ///                      "data/_html-template/",
 ///                      vec!["data/_html-template/_*"],
@@ -377,79 +377,4 @@ pub fn create_gitignore(proj: &MDBook) {
     debug!("[*]: Writing to .gitignore");
 
     f.write_all(&text.into_bytes()).expect("Could not write to file.");
-}
-
-// ------------------------------------------------------------------------------------------------
-// ------------------------------------------------------------------------------------------------
-
-// tests
-
-#[cfg(test)]
-mod tests {
-    extern crate tempdir;
-
-    use super::copy_files_except_ext;
-    use std::fs;
-
-    #[test]
-    fn copy_files_except_ext_test() {
-        let tmp = match tempdir::TempDir::new("") {
-            Ok(t) => t,
-            Err(_) => panic!("Could not create a temp dir"),
-        };
-
-        // Create a couple of files
-        if let Err(_) = fs::File::create(&tmp.path().join("file.txt")) {
-            panic!("Could not create file.txt")
-        }
-        if let Err(_) = fs::File::create(&tmp.path().join("file.md")) {
-            panic!("Could not create file.md")
-        }
-        if let Err(_) = fs::File::create(&tmp.path().join("file.png")) {
-            panic!("Could not create file.png")
-        }
-        if let Err(_) = fs::create_dir(&tmp.path().join("sub_dir")) {
-            panic!("Could not create sub_dir")
-        }
-        if let Err(_) = fs::File::create(&tmp.path().join("sub_dir/file.png")) {
-            panic!("Could not create sub_dir/file.png")
-        }
-        if let Err(_) = fs::create_dir(&tmp.path().join("sub_dir_exists")) {
-            panic!("Could not create sub_dir_exists")
-        }
-        if let Err(_) = fs::File::create(&tmp.path().join("sub_dir_exists/file.txt")) {
-            panic!("Could not create sub_dir_exists/file.txt")
-        }
-
-        // Create output dir
-        if let Err(_) = fs::create_dir(&tmp.path().join("output")) {
-            panic!("Could not create output")
-        }
-        if let Err(_) = fs::create_dir(&tmp.path().join("output/sub_dir_exists")) {
-            panic!("Could not create output/sub_dir_exists")
-        }
-
-        match copy_files_except_ext(&tmp.path(), &tmp.path().join("output"), true, &["md"]) {
-            Err(e) => panic!("Error while executing the function:\n{:?}", e),
-            Ok(_) => {},
-        }
-
-        // Check if the correct files where created
-        if !(&tmp.path().join("output/file.txt")).exists() {
-            panic!("output/file.txt should exist")
-        }
-        if (&tmp.path().join("output/file.md")).exists() {
-            panic!("output/file.md should not exist")
-        }
-        if !(&tmp.path().join("output/file.png")).exists() {
-            panic!("output/file.png should exist")
-        }
-        if !(&tmp.path().join("output/sub_dir/file.png")).exists() {
-            panic!("output/sub_dir/file.png should exist")
-        }
-        if !(&tmp.path().join("output/sub_dir_exists/file.txt")).exists() {
-            panic!("output/sub_dir/file.png should exist")
-        }
-
-    }
 }
